+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Next Release
+* Improve logging when can't find *.ps1 file ([#32](https://github.com/HuddleEng/octopose/issues/32))
 
 # v0.2.2
 * Fix local deploy not downloading correct packages when specifying version ([#30](https://github.com/HuddleEng/octopose/issues/30))