+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Next Release
-* Allow deploying of packages that have a different version to the release version ([#33](https://github.com/HuddleEng/octopose/issues/330))
+* Allow deploying of packages that have a different version to the release version ([#33](https://github.com/HuddleEng/octopose/issues/33))
 
 # v0.2.2
 * Fix local deploy not downloading correct packages when specifying version ([#30](https://github.com/HuddleEng/octopose/issues/30))