+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Next Release
+* Allow deploying of packages that have a different version to the release version ([#33](https://github.com/HuddleEng/octopose/issues/33))
 * Improve logging when can't find *.ps1 file ([#32](https://github.com/HuddleEng/octopose/issues/32))
 
 # v0.2.2