+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Next Release
+
+# v0.2.3
 * Allow deploying of packages that have a different version to the release version ([#33](https://github.com/HuddleEng/octopose/issues/33))
 * Improve logging when can't find *.ps1 file ([#32](https://github.com/HuddleEng/octopose/issues/32))
 * Only fail deploy for given project (and not all projects), if a package deploy step fails ([#34](https://github.com/HuddleEng/octopose/34))